backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
sqlalchemy==2.0.27
pydantic==2.6.1
python-multipart==0.0.9
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose==3.3.0
passlib==1.7.4